+++ tests/test_gerber_flow.py
@@ -185,6 +185,10 @@ class GerberFlowTestCase(unittest.TestCase):
             output_filename = tmp_file.name
         cnc_obj.export_gcode(output_filename)
         self.assertTrue(os.path.isfile(output_filename))
+        from gcode_safety import assert_safe_gcode
+        with open(output_filename, "r", encoding="utf-8", errors="replace") as fh:
+            exported = fh.read()
+        assert_safe_gcode(exported, cnc_obj.z_cut, cnc_obj.z_move)
         os.remove(output_filename)
 
         print(names)
